+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
-- Only run LaTeX once when converting
+- "Convert and download selected" button text is now sentence case rather than title case
 
 ### Fixed